Added plane intership

+++ course/draw.cpp
@@ -69,7 +69,7 @@ void DrawQt::drawmodel(vector<Point_3d> point, vector<int> r, vector<Point_3d> g
     vector<thread> threads;
 
     vector<Point_3d> new_point;
-    for (int i = 0; i < point.size(); i++) {
+    for (size_t i = 0; i < point.size(); i++) {
         vector<double> vec1 = {point[i].get_x(), point[i].get_y(), point[i].get_z(), 1};
         vector<double> p1 = camera.get_matrix() * vec1;
         new_point.push_back(Point_3d(p1[0], p1[1], p1[2]));
